Stories refactoring;
