+++ packages/grid/src/Grid.css
@@ -106,10 +106,6 @@
   --grid-groupHeader-fontWeight: var(--uitkGrid-groupHeader-fontWeight, var(--uitk-text-caption-fontWeight-strong));
 
   --grid-cell-padding: 0 var(--grid-padding);
-  --grid-rowHover-zIndex: 1;
-  --grid-cell-zIndex: 2;
-
-  --grid-columnSeparator-color: rgba(0, 0, 0, 0);
 
   --grid-background-primary: var(--uitkGrid-background-primary, var(--grid-background-primary-default));
   --grid-background-secondary: var(--uitkGrid-background-secondary, var(--grid-background-secondary-default));
@@ -132,7 +128,6 @@
   --grid-headerRowSeparator-color: var(--uitkGrid-headerRowSeparator-color, var(--grid-headerRowSeparator-color-default));
   --grid-headerRowSeparator-gap: var(--uitkGrid-headerRowSeparator-gap, 4px);
   --grid-headerRowSeparator-height: var(--uitkGrid-headerRowSeparator-height, 1px);
-  --grid-headerRowSeparator-width: var(--uitkGrid-headerRowSeparator-width, 1px);
   --grid-groupHeaderRowSeparator-color: var(--uitkGrid-groupHeaderRowSeparator-color, var(--grid-groupHeaderRowSeparator-color-default));
 
   --grid-groupHeader-color: var(--uitkGrid-groupHeader-color, var(--grid-groupHeader-color-default));
@@ -144,7 +139,7 @@
   --grid-cursor-borderColor: var(--uitkGrid-cursor-borderColor, var(--grid-cursor-borderColor-default));
   --grid-borderColor: var(--uitkGrid-borderColor, var(--grid-borderColor-default));
 
-  --grid-rowSeparator-width: var(--uitkGrid-rowSeparator-width, 1px);
+  --grid-rowSeparator-width: 1px;
   --grid-rowSeparator-color: var(--uitkGrid-rowSeparator-color, var(--grid-rowSeparator-color-default));
   --grid-columnSeparator-color: var(--uitkGrid-columnSeparator-color, var(--grid-columnSeparator-color-default));
   --grid-rowSeparator-color-divided: var(--uitkGrid-rowSeparator-color-divided, var(--grid-rowSeparator-color-divided-default));
