+++ packages/grid/src/HeaderCell.css
@@ -73,5 +73,4 @@
   width: 16px;
   background: transparent;
   cursor: col-resize;
-  z-index: 2;
 }
