+++ packages/grid/src/internal/ColumnDropTarget.css
@@ -4,5 +4,4 @@
   bottom: 0;
   width: var(--grid-columnDropTarget-width);
   background: var(--grid-columnDropTarget-color);
-  z-index: 3;
 }
