+++ packages/grid/src/internal/ColumnGhost.css
@@ -1,6 +1,5 @@
 .uitkGridColumnGhost {
   position: absolute;
-  z-index: 4;
   border: solid 1px var(--uitk-color-blue-500);
   box-shadow: 6px 6px 10px 10px rgba(0, 0, 0, 0.0967821);
   opacity: 0.85;
