+++ packages/grid/src/internal/Cursor.css
@@ -9,6 +9,5 @@
   right: 0;
   bottom: 0;
   left: 0;
-  z-index: 1;
   pointer-events: none;
 }
