+++ packages/grid/src/internal/TopLeftPart.css
@@ -2,7 +2,6 @@
   position: absolute;
   top: 0;
   left: 0;
-  z-index: 2;
 }
 
 .uitkGridTopLeftPart-raised {
